+++ Practical3/src/Data/jsonData.json
@@ -1,11 +1,10 @@
-[
-  {
+[{
     "id": 156235,
     "heading": "Shelby Mustang",
     "details": "Revolutionary thinking is at the heart of every idea from Automobili Lamborghini. Whether it is aerospace-inspired design or technologies applied to the naturally aspirated V12 engine",
     "image": "https://hips.hearstapps.com/hmg-prod/images/2022-ford-mustang-shelby-gt500-02-1636734552.jpg?crop=1.00xw:0.891xh;0,0.0759xh&resize=2048:*",
     "specs": "3.8-litre Whipple supercharger to the 5.2-litre V8, 900 horsepower on 93-octane premium gas.",
-    "price": " 110000000"
+    "price": "110000000"
   },
   {
     "id": 24653,
