--- Practical3/src/components/NavagationBar.vue
@@ -1,200 +1,0 @@
-<template>
-  <div class="Navagation_bar">
-    <div>Navagation bar from navagation</div>
-    <div>
-      <button
-        type="button"
-        class="btn btn-primary"
-        data-toggle="modal"
-        data-target="#AddCarForm"
-      >
-        Add Car
-      </button>
-      <div
-        class="modal fade"
-        id="AddCarForm"
-        tabindex="-1"
-        role="dialog"
-        aria-labelledby="AddCarFormLabel"
-        aria-hidden="true"
-      >
-        <div class="modal-dialog" role="document">
-          <div class="modal-content">
-            <div class="modal-header">
-              <h5 class="modal-title" id="AddCarFormLabel">Add Car</h5>
-              <button
-                type="button"
-                class="close"
-                data-dismiss="modal"
-                aria-label="Close"
-              >
-                <span aria-hidden="true">&times;</span>
-              </button>
-            </div>
-
-            <div class="modal-body">
-              <Form @submit="onSubmit" :validation-schema="schema">
-                <div>
-                  <Field
-                    id="heading"
-                    name="heading"
-                    type="text"
-                    placeholder="Heading:- Ex:- Ambassador"
-                    class="form-control"
-                  />
-                  <ErrorMessage name="heading" class="text-danger h6" />
-                </div>
-
-                <div>
-                  <Field
-                    id="details"
-                    name="details"
-                    type="text"
-                    placeholder="Description Ex:- Owner of this Organization"
-                    class="form-control"
-                  />
-                  <ErrorMessage name="details" class="text-danger h6" />
-                </div>
-
-                <div>
-                  <Field
-                    id="specs"
-                    name="specs"
-                    type="text"
-                    placeholder="Specification Ex:- Power-HP/500CC Petrol/Diesel"
-                    class="form-control"
-                  />
-                  <ErrorMessage name="specs" class="text-danger h6" />
-                </div>
-
-                <div>
-                  <Field
-                    id="image"
-                    name="image"
-                    type="text"
-                    placeholder="Image Link Ex:- https://abc.com"
-                    class="form-control"
-                  />
-                  <ErrorMessage name="image" class="text-danger h6" />
-                </div>
-
-                <div>
-                  <Field
-                    id="price"
-                    name="price"
-                    type="number"
-                    placeholder="Ex:- 123456 Only/-"
-                    class="form-control"
-                  />
-                  <ErrorMessage name="price" class="text-danger h6" />
-                </div>
-
-                <button class="btn btn-secondary">Submit</button>
-              </Form>
-            </div>
-
-            
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script>
-import { Field, Form, ErrorMessage } from "vee-validate";
-import * as yup from "yup";
-export default {
-  name: "NavagationBar",
-  components: {
-    Field,
-    Form,
-    ErrorMessage,
-  },
-  props: {
-    upcomingData: Object,
-  },
-  methods: {
-    onSubmit(values) {
-      // this.carsDetails.push = values;
-      alert(JSON.stringify(values, null, 2));
-      console.log("form values: ", values);
-      this.$emit("onFormSubmitHandler", JSON.stringify(values));
-    },
-  },
-  setup() {
-    const schema = yup.object().shape({
-      heading: yup.string().required("Car heading is Required!"),
-      details: yup
-        .string()
-        .required("Car Description is Required!")
-        .min(30)
-        .max(120),
-      specs: yup
-        .string()
-        .required("Car Specification is Required!")
-        .min(12)
-        .max(30),
-      image: yup
-        .string()
-        .required("Car Image is Required!")
-        .url("Car Image must be url Format!"),
-      price: yup.number().required("Car Price is Required!"),
-    });
-
-    return {
-      schema,
-    };
-  },
-};
-</script>
-
-<style>
-.modal-body input {
-  border-radius: 15px;
-  background-color: rgb(222, 222, 222);
-  border: 1px solid gray;
-  padding: 5px;
-  margin: 6px;
-}
-.Navagation_bar {
-  min-width: 80%;
-  background-color: rgb(84, 121, 241);
-  height: 55px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: x-large;
-  font-weight: 600;
-  border-radius: 50px;
-  padding: 0 80px;
-}
-.add_button {
-  background-color: green;
-  border-radius: 25px;
-  padding: 8px 15px;
-  font-size: 21px;
-  color: white;
-  cursor: pointer;
-}
-.add_button:hover {
-  background-color: green;
-  border-radius: 25px;
-  border-top: 2px solid darkgreen;
-  border-top: 2px solid darkgreen;
-  padding: 8px 15px;
-  font-size: 21px;
-  color: white;
-  cursor: pointer;
-}
-.add_button:active {
-  background-color: green;
-  border-radius: 25px;
-  border-top: 1px solid darkgreen;
-  border-top: 1px solid darkgreen;
-  padding: 8px 15px;
-  font-size: 21px;
-  color: white;
-  cursor: pointer;
-}
-</style>
